[C++ SDK] Fixed OSS executor build
