+++ ydb/public/sdk/cpp/include/ydb-cpp-sdk/client/types/executor/executor.h
@@ -5,6 +5,7 @@
 #endif
 
 #include <functional>
+#include <memory>
 #include <mutex>
 
 namespace NYdb::inline Dev {
