+++ ydb/public/sdk/cpp/src/client/impl/executor/ya.make
@@ -1,0 +1,7 @@
+LIBRARY()
+
+SRCS(
+    executor_impl.cpp
+)
+
+END()
