+++ ydb/public/sdk/cpp/src/client/types/executor/ya.make
@@ -5,6 +5,7 @@ SRCS(
 )
 
 PEERDIR(
+    ydb/public/sdk/cpp/src/client/impl/executor
     ydb/public/sdk/cpp/src/client/impl/internal/thread_pool
     ydb/public/sdk/cpp/src/client/types/exceptions
 )
